+++ db/migrate/20200226031655_create_trades.rb
@@ -3,7 +3,7 @@ class CreateTrades < ActiveRecord::Migration[6.0]
     create_table :trades do |t|
       t.datetime :purchase_date, null: false
       t.integer :quantity, null: false
-      t.integer :price, null: false
+      t.decimal :price, null: false
       t.integer :user_id, null: false
       t.timestamps
     end
